requirements_extended.txt
asgiref==3.7.2
certifi==2023.5.7
charset-normalizer==3.1.0
dj-database-url==2.0.0
Django==4.2.2
django-cors-headers==4.1.0
python-dotenv==1.0.0
gunicorn==20.1.0
idna==3.4
Pillow==9.5.0
requests==2.31.0
sqlparse==0.4.4
stripe==5.4.0
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.3
whitenoise==6.5.0

# New dependencies for improvements
django-redis==5.3.0
redis==4.6.0
celery==5.3.1
reportlab==4.0.4
django-import-export==3.2.0
django-allauth==0.54.0
Pillow==9.5.0
django-imagekit==4.1.0
django-storages==1.13.2
boto3==1.28.25
